Separate Updater class in its own folder and separate input parsing logic in classes
